adding some good functionalities

+++ front_end/src/components/ConfirmAlert.jsx
@@ -7,18 +7,21 @@ import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
 import axios from 'axios';
 
-function ConfirmAlert({openAlert, setOpenAlert, content}) {
+function ConfirmAlert({openAlert, setOpenAlert, content, titleAndDescription}) {
   const handleClose = () => {
     setOpenAlert(false);
   };
 
   function handleDeletionConfirm() {
-    axios.delete("http://localhost:8080/api/products/"+ content.product.id)
-    .then(() => {
-        console.log("product with id "+ content.product.id +" is deleted")
-        setOpenAlert(false)
-    })
-    .catch((error) => console.log("errot => "+ error))
+    if(titleAndDescription.action === "delete") {
+      axios.delete("http://localhost:8080/api/products/"+ content.product.id)
+      .then(() => {
+          console.log("product with id "+ content.product.id +" is deleted")
+          setOpenAlert(false)
+      })
+      .catch((error) => console.log("errot => "+ error))
+    }
+    else setOpenAlert(false)
   }
   
 
@@ -31,17 +34,20 @@ function ConfirmAlert({openAlert, setOpenAlert, content}) {
         aria-describedby="alert-dialog-description"
       >
         <DialogTitle id="alert-dialog-title">
-          {"Are you sure you want to delete this Product?"}
+          {/* {"Are you sure you want to delete this Product?"} */}
+          {titleAndDescription.title}
         </DialogTitle>
         <DialogContent>
           <DialogContentText id="alert-dialog-description">
-            After deletion there is no going back it will be removed for ever 
+            {/* After deletion there is no going back it will be removed for ever  */}
+            {titleAndDescription.description}
           </DialogContentText>
         </DialogContent>
         <DialogActions>
           <Button onClick={handleClose}>Cancel</Button>
-          <Button onClick={handleDeletionConfirm}  style={{color: "red"}}>
-            Delete
+          <Button onClick={handleDeletionConfirm}  style={{color: titleAndDescription.color}}>
+            {/* Delete */}
+            {titleAndDescription.action}
           </Button>
         </DialogActions>
       </Dialog>
